move tests to /unit/ folder
